refactor(manage): 题库模块增加返回信息
- 题目模块增加获取题目数量功能
- 修改题目业务名称 ExamQuestion -> Question

+++ eduflex-manage/src/main/java/com/eduflex/manage/domain/vo/RepoVo.java
@@ -8,13 +8,13 @@ public class RepoVo extends Repo {
     private String courseName;
 
     // 单选题数量
-    private Integer singleChoiceCount;
+    private Long singleChoiceCount;
     // 多选题数量
-    private Integer multipleChoiceCount;
+    private Long multipleChoiceCount;
     // 判断题数量
-    private Integer judgeCount;
+    private Long judgeCount;
     // 填空题数量
-    private Integer blankCount;
+    private Long blankCount;
     // 简答题数量
-    private Integer shortAnswerCount;
+    private Long shortAnswerCount;
 }
